refactor: split report tests by behavior
